requirements.txt
langchain_core
langchain_community
langchain-openai
streamlit
python-dotenv
